refactor(frontend): rename ETH send components
