+++ src/frontend/src/eth/components/send/EthSendModal.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-	import SendTokenModal from '$eth/components/send/SendTokenModal.svelte';
+	import EthSendTokenModal from '$eth/components/send/EthSendTokenModal.svelte';
 	import SendTokenContext from '$eth/components/send/SendTokenContext.svelte';
 	import { token } from '$lib/stores/token.store';
 </script>
 
 <SendTokenContext token={$token}>
-	<SendTokenModal on:nnsClose />
+	<EthSendTokenModal on:nnsClose />
 </SendTokenContext>
