+++ src/frontend/src/eth/components/send/Send.svelte
@@ -2,7 +2,7 @@
 	import { modalStore } from '$lib/stores/modal.store';
 	import { addressNotLoaded } from '$lib/derived/address.derived';
 	import { modalSend } from '$lib/derived/modal.derived';
-	import SendModal from '$eth/components/send/SendModal.svelte';
+	import EthSendModal from '$eth/components/send/EthSendModal.svelte';
 	import { waitWalletReady } from '$lib/services/actions.services';
 	import SendButton from '$lib/components/send/SendButton.svelte';
 	import { loadTokenAndRun } from '$icp/services/token.services';
@@ -32,5 +32,5 @@
 <SendButton on:click={async () => await openSend()} {compact} />
 
 {#if $modalSend && $modalStore?.data === modalId}
-	<SendModal on:nnsClose />
+	<EthSendModal on:nnsClose />
 {/if}
